tests/requirements-test.txt
# Testing dependencies for MCP Zoo

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0

# API response caching
vcrpy>=5.0.0
vcrpy-unittest>=1.0.0

# Mocking and fixtures
responses>=0.23.0
freezegun>=1.2.0
faker>=19.0.0

# Type checking (optional)
mypy>=1.5.0
types-requests>=2.31.0

# Code quality
pytest-xdist>=3.3.0  # Parallel test execution
pytest-html>=3.2.0  # HTML test reports

# MCP SDK for testing
mcp>=0.1.0

# Utilities
python-dotenv>=1.0.0

